+++ VersionHistory.md
@@ -8,6 +8,7 @@ is updated in preparation for publishing an updated NuGet package.
 Prefix the description of the change with `[major]`, `[minor]` or `[patch]` in accordance with [SemVer](http://semver.org).
 
 * [patch] Accept json responses without Content-Length set.
+* [patch] Detach `WebServiceResponse.Content` to allow it to be read as a stream.
 
 ## Released